Fix bug: exporting data to .dat results in loss of one datapoint
Exporting data with non-integer us tres to a .dat file resulted
in the loss of the last raw uninterpolated datapoint.
When exporting multiple files to a concatenated .dat file, the last
datapoint of each would be missing in the concatenated .dat file,
and generally the data for each file was shifted one data point to
the left. In practice, this had no effect on results, because
the concatenated .dat is only used by kilosort, which is then
always followed by manual curation.

This was due to another floating-point roundoff error, and was
fixed by rounding to the nearest integer multiple of tres
instead of truncating, when calculating dt0i in MultiStream.__call__